Refactor database (all id = long, correct PKeys, FKeys, setup many to many entities)

+++ backend/LeetWars.Core/LeetWars.Core.Common/DTO/Challenge/ChallengeDto.cs
@@ -7,7 +7,7 @@ namespace LeetWars.Core.Common.DTO.Challenge
     {
         public string Title { get; set; } = string.Empty;
         public string Instructions { get; set; } = string.Empty;
-        public int LevelId { get; set; }
+        public long LevelId { get; set; }
         public ChallengeCategory Category { get; set; }
 
     }
