+++ backend/LeetWars.Core/LeetWars.Core.Common/DTO/Challenge/NewChallengeDto.cs
@@ -9,7 +9,7 @@ namespace LeetWars.Core.DAL.Entities
         public long? CreatedBy { get; set; }
         public string Title { get; set; } = string.Empty;
         public string Instructions { get; set; } = string.Empty;
-        public int LevelId { get; set; }
+        public long LevelId { get; set; }
         public ChallengeCategory Category { get; set; }
         public ICollection<TagDto>? Tags { get; set; }
         public ICollection<NewChallengeVersionDto>? Versions { get; set; }
