+++ backend/LeetWars.Core/LeetWars.Core.Common/DTO/Language/LanguageDto.cs
@@ -2,7 +2,7 @@
 
 namespace LeetWars.Core.Common.DTO.Language
 {
-    public class LanguageDto : Entity<int>
+    public class LanguageDto : Entity<long>
     {
         public string Name { get; set; } = string.Empty;
     }
