+++ backend/LeetWars.Core/LeetWars.Core.DAL/Entities/Tag.cs
@@ -1,6 +1,6 @@
 ﻿namespace LeetWars.Core.DAL.Entities
 {
-    public class Tag : Entity<int>
+    public class Tag : Entity<long>
     {
         public string Name { get; set; }
         public ICollection<Challenge> Challenges { get; } = new List<Challenge>();
